Modify round to next plugin
